Rename files and fix missing update
